refactor: make PostPreview use ImageLink component
